feat: Add Event file and remove unnecessary annotaion
