+++ ProjectManager/ProjectManager/ViewModel/DetailViewModel.swift
